+++ src/ml_lifecycle_platform/pipeline/featurize.py
@@ -15,7 +15,7 @@ from ml_lifecycle_platform.common.constants import (
     TEST_CSV,
     TRAIN_CSV,
 )
-from ml_lifecycle_platform.common.mlflow_utils import ensure_experiment
+from ml_lifecycle_platform.runtime.mlflow import ensure_experiment
 from ml_lifecycle_platform.runtime.bootstrap import get_runtime_context
 from ml_lifecycle_platform.core.batch_contracts import validate_labeled_dataset
 from ml_lifecycle_platform.core.model_specs import load_model_spec
