+++ src/ml_lifecycle_platform/serving/model_store.py
@@ -13,7 +13,7 @@ except Exception:  # pragma: no cover
     mlflow = None  # type: ignore[assignment]
     MlflowException = Exception  # type: ignore[assignment, misc]
 
-from ml_lifecycle_platform.common.mlflow_utils import client as get_mlflow_client
+from ml_lifecycle_platform.runtime.mlflow import client as get_mlflow_client
 from ml_lifecycle_platform.runtime.bootstrap import configure_mlflow
 
 from .constants import ALIAS_CANDIDATE, ALIAS_PROD
